refactor(frontend): extract paste-to-terminal modal from main.ts (#312)
Continues the conservative split: pull the paste modal (clipboard / manual
text → active terminal) into paste.ts (~180 lines), with its MAX_PASTE_CHARS
cap and pasteOpener state.

getActiveTerminal (a 4-line core helper reading currentActiveTabId /
currentTerminals, also used by the actions menu + file-attach) stays in
main.ts and is exported for paste.ts. closePasteModal (Escape handler) and
runPasteFlow (actions menu) are exported from paste.ts. Same TDZ-safe
pattern; the shared pasteModal const stays in main.ts for the modal-open
guard.

main.ts: 3651 → 3486. tsc + Biome clean, vite build green, frontend tests
pass (66).
